server/app/requirements.txt
Flask==1.1.2
PyMySQL==0.10.1
python-dotenv
openai
flask
requests
click==7.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
shortuuid==1.0.1
Werkzeug==1.0.1